Neutron Nova Interactions Support
Configure Notification System Options as part of a new class:
neutron::server::notifications.

This class is part of neutron::server because the notifications are sent
by Neutron API service which is configured in neutron-server nodes.

Also add a new type/provider to translate the nova admin tenant to an ID
and set it in neutron.conf.

Closes-bug #1306694
Change-Id: I94334e92bc0057665a281763a4543b127fde7f72
Signed-off-by: Emilien Macchi <emilien.macchi@enovance.com>